+++ pyphyschemtools/__init__.py
@@ -3,8 +3,8 @@
 The pyphyschemtools library provides a comprehensive suite of utilities for Physical Chemistry, ranging from spectroscopic unit management to kinetic modeling and cheminformatics.
 """
 
-__version__ = "0.8.14"
-__last_update__ = "2026-03-31"
+__version__ = "0.9.0"
+__last_update__ = "2026-04-01"
 
 import importlib
 import importlib.util
@@ -34,7 +34,7 @@ def __getattr__(name):
         ".ML", ".PeriodicTable", ".Chem3D",  
         ".aithermo", ".cheminformatics", ".kinetics", 
         ".misc", ".nano", ".orgachem",
-        ".spectra", ".survey", 
+        ".spectro", ".survey", 
         ".sympyUtilities", ".tools4AS", ".units"
     ]